Add useScopedColumns
